@@ -12,49 +13,55 @@
     "i18n-iso-countries": "^3.7.8",
     "immutability-helper": "^3.0.0",
     "js-yaml": "^3.12.0",
-    "jsonwebtoken": "^8.5.1",
     "moment": "^2.23.0",
     "moment-timezone": "^0.5.23",
-    "react": "^16.4.2",
+    "postcss": "^7.0.23",
+    "postcss-cli": "^6.1.3",
+    "postcss-loader": "^3.0.0",
+    "react": "16.10",
     "react-date-picker": "^7.0.0",
-    "react-dom": "^16.4.2",
+    "react-dom": "^16.6.3",
     "react-fade-in": "^0.1.6",
     "react-ga": "^2.6.0",
     "react-google-recaptcha": "^1.0.4",
-    "react-grid-layout": "^0.16.6",
+    "react-grid-layout": "^0.17.1",
     "react-intl": "^2.8.0",
     "react-maskinput": "^1.0.1",
-    "react-redux": "^5.0.7",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
-    "react-scripts-ts": "^2.17.0",
-    "redux": "^4.0.0",
+    "react-redux": "^6.0.0",
+    "react-router": "^5.1.2",
+    "react-router-dom": "^5.1.2",
+    "react-scripts": "3.0.1",
+    "recharts": "^1.8.5",
+    "redux": "^4.0.1",
     "redux-saga": "^0.16.2",
     "sinon": "^7.2.3",
-    "tslint-eslint-rules": "^5.4.0",
+    "tslint": "^5.13.1",
     "ua-parser-js": "^0.7.19",
-    "webpack": "^3.11.0"
+    "ws": "^7.2.0"
   },
   "scripts": {
     "build:css": "postcss src/index.pcss -o src/index.css",
     "watch:css": "postcss src/index.pcss -o src/index.css -w",
     "lint": "tsc --noEmit && tslint --project .",
     "lint:fix": "tslint --project . --fix",
+    "mockAPI": "node mocks/mockAPI.js",
+    "mockWS": "node mocks/mockWS.js",
     "start": "npm-run-all -p watch:css start-js",
-    "start-js": "react-app-rewired start --scripts-version react-scripts-ts start",
-    "build-js": "react-app-rewired build --scripts-version react-scripts-ts build",
+    "start-mock": "npm-run-all -p watch:css start-js mockAPI mockWS",
+    "start-js": "react-scripts start",
+    "build-js": "react-scripts build",
     "build": "npm-run-all build:css build-js clean:mapfiles",
     "clean:mapfiles": "find build -name '*.map' -exec rm {} ';'",
-    "test": "jest",
-    "test:coverage": "jest --collectCoverage=true",
-    "test:ci": "jest --ci",
+    "test": "react-scripts test",
+    "test:coverage": "react-scripts test --collectCoverage=true",
+    "test:ci": "react-scripts test --ci",
     "spec:integration": "cypress open",
-    "eject": "react-scripts-ts eject",
+    "eject": "react-scripts eject",
     "clean": "rimraf build",
     "analyze": "source-map-explorer build/static/js/main.*"
   },
   "devDependencies": {
-    "@openware/coding-standards": "0.1.6",
+    "@openware/coding-standards": "2.3.0",
     "@types/axios": "^0.14.0",
     "@types/classnames": "^2.2.6",
     "@types/enzyme": "^3.1.13",
@@ -103,24 +107,29 @@
     "source-map-explorer": "^1.7.0",
     "ts-jest": "^23.1.3",
     "ts-node": "^7.0.1",
-    "typescript": "^3.0.1",
+    "typescript": "3.4.5",
+    "webpack": "4.29.6",
     "webpack-obfuscator": "^0.18.0"
   },
   "husky": {
     "hooks": {
       "commit-msg": "./scripts/validate_commit.js $HUSKY_GIT_PARAMS && (cd \"./$GIT_PREFIX\"; yarn lint)",
-      "pre-push": "yarn lint && yarn test"
+      "pre-push": "yarn lint && CI=true yarn test --no-watch"
     }
   },
   "prettier": {
-    "trailingComma": "es5",
+    "trailingComma": "es6",
     "tabWidth": 4,
     "printWidth": 130,
     "semi": true,
     "singleQuote": true,
     "tslintIntegration": true,
-    "eslintIntegration": true,
     "stylelintIntegration": true,
     "ignorePath": ".gitignore"
-  }
+  },
+  "browserslist": [
+    ">0.2%",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
