@@ -7,7 +7,7 @@ export const defaultConfig: Config = {
         tradeUrl: '',
         applogicUrl: '',
         rangerUrl: '',
-        tenkoUrl: '',
+        arkeUrl: '',
     },
     minutesUntilAutoLogout: '5',
     rangerReconnectPeriod: '1',
@@ -19,7 +19,6 @@ export const defaultConfig: Config = {
     storage: {},
     gaTrackerKey: '',
     msAlertDisplayTime: '5000',
-    licenseKey: '',
     incrementalOrderBook: false,
 };
 
@@ -38,10 +37,10 @@ Cryptobase.config = { ...window.env };
 Cryptobase.config.storage = Cryptobase.config.storage || {};
 Cryptobase.config.captcha = Cryptobase.config.captcha || defaultConfig.captcha;
 
-export const authUrl = () => Cryptobase.config.api.authUrl;
 export const tradeUrl = () => Cryptobase.config.api.tradeUrl;
+export const arkeUrl = () => Cryptobase.config.api.arkeUrl || tradeUrl();
+export const authUrl = () => Cryptobase.config.api.authUrl;
 export const applogicUrl = () => Cryptobase.config.api.applogicUrl;
-export const tenkoUrl = () => Cryptobase.config.api.tenkoUrl;
 export const rangerUrl = () => Cryptobase.config.api.rangerUrl;
 export const minutesUntilAutoLogout = (): string => Cryptobase.config.minutesUntilAutoLogout || '5';
 export const withCredentials = () => Cryptobase.config.withCredentials;
