@@ -1,7 +1,7 @@
 import classnames from 'classnames';
 import * as React from 'react';
 
-interface SummaryFieldProps {
+export interface SummaryFieldProps {
     /**
      * Additional class name for styling. By default element receives `cr-input` class
      * @default empty
@@ -12,11 +12,6 @@ interface SummaryFieldProps {
      */
     message: string;
     /**
-     * Border item for summary field. It can be 'rectangle', 'circle' or 'empty-circle'
-     * @default 'rectangle'
-     */
-    borderItem?: string;
-    /**
      * Content will be displayed instead of amount and currency, if it is necessary
      */
     content: JSX.Element;
@@ -25,16 +20,11 @@ interface SummaryFieldProps {
 /**
  * Component to display currency amount with specific label.
  */
-const SummaryField: React.FunctionComponent<SummaryFieldProps> = props => {
-    const { message, className, borderItem, content } = props;
-    const cx = classnames('cr-summary-field', {
-            'cr-summary-field-border-item': borderItem,
-        },
-        className);
-    const symbol = `cr-summary-field-border-item-${borderItem}`;
+export const SummaryField: React.FunctionComponent<SummaryFieldProps> = props => {
+    const { message, className, content } = props;
+    const cx = classnames('cr-summary-field', className);
     return (
         <div className={cx}>
-            <div className={symbol} />
             <span className="cr-summary-field-message">
             {message}
           </span>
@@ -43,9 +33,4 @@ const SummaryField: React.FunctionComponent<SummaryFieldProps> = props => {
           </span>
         </div>
     );
-};
-
-export {
-    SummaryField,
-    SummaryFieldProps,
 };
