@@ -1,9 +1,11 @@
 import Axios from 'axios';
 import MockAdapter from 'axios-mock-adapter';
-import WebSocket = require('isomorphic-ws');
 import { Action, Middleware } from 'redux';
 import configureMockStore from 'redux-mock-store';
 import { Config, Cryptobase } from '../api';
+
+// tslint:disable-next-line
+import * as WebSocket from 'ws';
 
 const mockConfig: Config = {
     api: {
@@ -11,7 +13,7 @@ const mockConfig: Config = {
         tradeUrl: '/api/v2/peatio',
         applogicUrl: '/api/v2/applogic',
         rangerUrl: '/api/v2/ranger',
-        tenkoUrl: '/api/v2/tenko',
+        arkeUrl: '/api/v2/arke',
     },
     minutesUntilAutoLogout: '5',
     rangerReconnectPeriod: '1',
@@ -22,7 +24,6 @@ const mockConfig: Config = {
         siteKey: '',
     },
     gaTrackerKey: '',
-    licenseKey: '',
     incrementalOrderBook: false,
 };
 
