@@ -6,6 +6,7 @@ import { gridLayoutReducer } from './public/gridLayout/reducer';
 import { changeLanguageReducer  } from './public/i18n';
 import { klineReducer  } from './public/kline';
 import { marketsReducer } from './public/markets';
+import { memberLevelsReducer } from './public/memberLevels';
 import {
     depthReducer,
     incrementDepthReducer,
@@ -16,8 +17,8 @@ import { recentTradesReducer  } from './public/recentTrades';
 import { apiKeysReducer } from './user/apiKeys';
 import { authReducer  } from './user/auth';
 import { beneficiariesReducer } from './user/beneficiaries';
+import { getGeetestCaptchaReducer } from './user/captcha';
 import { sendEmailVerificationReducer } from './user/emailVerification';
-import { guardReducer } from './user/guard';
 import { historyReducer  } from './user/history';
 import {
     documentsReducer,
@@ -48,6 +49,7 @@ export const publicReducer = combineReducers({
     kline: klineReducer,
     alerts: alertReducer,
     rgl: gridLayoutReducer,
+    memberLevels: memberLevelsReducer,
 });
 
 export const userReducer = combineReducers({
@@ -68,6 +70,6 @@ export const userReducer = combineReducers({
     ordersHistory: ordersHistoryReducer,
     openOrders: openOrdersReducer,
     sendEmailVerification: sendEmailVerificationReducer,
+    captchaKeys: getGeetestCaptchaReducer,
     withdrawLimit: withdrawLimitReducer,
-    guard: guardReducer,
 });
