@@ -1,17 +1,23 @@
 import { CommonError } from '../../types';
 import {
-    AUTH_ERROR,
     AUTH_LOGOUT_FAILURE,
     AUTH_LOGOUT_FETCH,
     AUTH_SIGN_IN_ERROR,
     AUTH_SIGN_IN_FETCH,
     AUTH_SIGN_IN_REQUIRE_2FA,
     AUTH_SIGN_UP_FETCH,
+    AUTH_SIGN_UP_ERROR,
     AUTH_SIGN_UP_REQUIRE_VERIFICATION,
     AUTH_TEST_STATE,
     AUTH_VERIFICATION_FETCH,
     AUTH_VERIFICATION_SUCCESS,
 } from './constants';
+
+export interface GeetestCaptchaResponse {
+    geetest_challenge: string;
+    geetest_validate: string;
+    geetest_seccode: string;
+}
 
 export interface SignInFetch {
     type: typeof AUTH_SIGN_IN_FETCH;
@@ -40,13 +46,13 @@ export interface SignUpFetch {
     payload: {
         email: string;
         password: string;
-        recaptcha_response?: string;
-        refId?: string;
+        captcha_response?: string | GeetestCaptchaResponse;
+        refid?: string;
     };
 }
 
 export interface SignUpError {
-    type: typeof AUTH_ERROR;
+    type: typeof AUTH_SIGN_UP_ERROR;
     payload: CommonError;
 }
 
@@ -115,7 +121,7 @@ export const signUp = (payload: SignUpFetch['payload']): SignUpFetch => ({
 });
 
 export const signUpError = (payload: SignUpError['payload']): SignUpError => ({
-    type: AUTH_ERROR,
+    type: AUTH_SIGN_UP_ERROR,
     payload,
 });
 
