@@ -2,7 +2,12 @@
 import { call, put } from 'redux-saga/effects';
 import { API, RequestOptions } from '../../../../../api';
 import { alertPush } from '../../../../public/alert';
-import { sendCodeData, sendCodeError, SendCodeFetch } from '../actions';
+import {
+    sendCodeData,
+    sendCodeError,
+    SendCodeFetch,
+    resendCode,
+} from '../actions';
 
 const sessionsConfig: RequestOptions = {
     apiVersion: 'barong',
@@ -14,7 +19,12 @@ export function* sendCodeSaga(action: SendCodeFetch) {
         yield put(sendCodeData());
         yield put(alertPush({message: ['success.phone.verification.send'], type: 'success'}));
     } catch (error) {
-        yield put(sendCodeError(error));
-        yield put(alertPush({message: error.message, code: error.code, type: 'error'}));
+        if (error.message.toString() === 'resource.phone.exists') {
+            yield put(resendCode(action.payload));
+            yield put(alertPush({message: error.message, type: 'success'}));
+        } else {
+            yield put(sendCodeError(error));
+            yield put(alertPush({message: error.message, code: error.code, type: 'error'}));
+        }
     }
 }
