@@ -156,6 +156,10 @@ export const en = {
     'page.body.wallets.beneficiaries.tipName': 'Name',
     'page.body.wallets.beneficiaries.tipDescription': 'Note',
 
+    'page.body.wallets.beneficiaries.failAddModal.header': 'Warning',
+    'page.body.wallets.beneficiaries.failAddModal.content': 'You need to confirm your account to add a beneficiary',
+    'page.body.wallets.beneficiaries.failAddModal.button': 'Confirm an account',
+
     'page.body.openOrders.tab.all': 'All',
     'page.body.openOrders.tab.open': 'Open',
     'page.body.openOrders.header.orderType': 'Order Type',
@@ -207,6 +211,7 @@ export const en = {
     'page.body.history.withdraw.content.status.succeed': 'Succeed',
     'page.body.history.withdraw.content.status.failed': 'Failed',
     'page.body.history.withdraw.content.status.confirming': 'Confirming',
+    'page.body.history.withdraw.content.status.errored': 'Error',
 
     'page.body.history.trade': 'Trade History',
     'page.body.history.trade.header.id': 'ID',
@@ -219,6 +224,7 @@ export const en = {
     'page.body.history.trade.header.total': 'Total',
     'page.body.history.trade.header.amount': 'Amount',
     'page.body.history.trade.header.balance': 'Balance',
+
 
     'page.body.profile.header.account': 'Profile',
 
@@ -432,7 +438,7 @@ export const en = {
     'resource.api_key.invalid_otp': 'OTP code is invalid',
     'resource.phone.twillio': 'Something wrong with Twilio Client',
     'resource.phone.invalid_num': 'Phone number is invalid',
-    'resource.phone.exists': 'Phone number already exists',
+    'resource.phone.exists': 'Phone number was not verified. Verification code was resent.',
     'resource.phone.number_exist': 'Phone number already exists',
     'resource.phone.verification_invalid': 'Phone is not found or verification code is invalid',
     'resource.documents.limit_reached': 'Maximum number of documents already reached',
